+++ src/app/lista-voluntarios/lista-voluntarios.component.ts
@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { VoluntariosService } from 'projects/entities/src/lib/voluntarios.service';
 import { Voluntario } from 'projects/entities/src/lib/voluntarios/voluntario';
+import { ConfirmaRemocaoComponent } from '../confirma-remocao/confirma-remocao.component';
+import { MatDialog } from '@angular/material';
+import { DocumentChangeAction } from '@angular/fire/firestore';
 
 @Component({
   selector: 'app-lista-voluntarios',
@@ -12,13 +15,14 @@ export class ListaVoluntariosComponent implements OnInit {
 
   ageValue: number = 0;
   searchValue: string = '';
-  voluntarios: Array<Voluntario>;
+  voluntariosDocRef: Array<DocumentChangeAction<Voluntario>>;
   age_filtered_items: Array<any>;
   name_filtered_items: Array<any>;
 
   constructor(
     public service: VoluntariosService,
-    private router: Router
+    private router: Router,
+    private dialog: MatDialog
   ) { }
 
   ngOnInit() {
@@ -27,34 +31,42 @@ export class ListaVoluntariosComponent implements OnInit {
 
   getData() {
     this.service.getVoluntarios()
-      .subscribe((result: Voluntario[]) => {
-        this.voluntarios = result;
+      .subscribe((result: DocumentChangeAction<Voluntario>[]) => {
+        this.voluntariosDocRef = result;
         this.age_filtered_items = result;
         this.name_filtered_items = result;
       })
   }
 
-  viewDetails(voluntario) {
+  viewDetails(voluntario: DocumentChangeAction<Voluntario>) {
     this.router.navigate(['/details/' + voluntario.payload.doc.id]);
   }
 
-  editRules(voluntario){
-    this.router.navigate(['/rules/' + voluntario.payload.doc.id]);
+  editRoles(voluntario: DocumentChangeAction<Voluntario>) {
+    this.router.navigate(['/roles/' + voluntario.payload.doc.id]);
   }
 
-  delete(voluntario){
-    this.service.deleteVoluntario(voluntario.payload.doc.id)
-      .then(
-        res => {
-          this.router.navigate(['/']);
-        },
-        err => {
-          this.showErrorMessage(err);
-        }
-      )
+  delete(voluntario: DocumentChangeAction<Voluntario>) {
+    const dialogRef = this.dialog.open(ConfirmaRemocaoComponent, {
+      width: '250px',
+      data: voluntario
+    });
+
+    dialogRef.afterClosed().subscribe((result:boolean) => {
+      console.log('The dialog was closed: '+result);
+      if (result) this.service.deleteVoluntario(voluntario.payload.doc.id)
+        .then(
+          res => {
+            this.router.navigate(['/']);
+          },
+          err => {
+            this.showErrorMessage(err);
+          }
+        )
+    });
   }
 
-  showErrorMessage(err){
+  showErrorMessage(err) {
     console.log(err);
   }
 
@@ -67,7 +79,7 @@ export class ListaVoluntariosComponent implements OnInit {
     this.service.searchVoluntario(value)
       .subscribe(result => {
         this.name_filtered_items = result;
-        this.voluntarios = this.combineLists(result, this.age_filtered_items);
+        this.voluntariosDocRef = this.combineLists(result, this.age_filtered_items);
       });
   }
 
@@ -75,7 +87,7 @@ export class ListaVoluntariosComponent implements OnInit {
     this.service.searchVoluntarioPelaIdade(event.value)
       .subscribe(result => {
         this.age_filtered_items = result;
-        this.voluntarios = this.combineLists(result, this.name_filtered_items);
+        this.voluntariosDocRef = this.combineLists(result, this.name_filtered_items);
       })
   }
